New required compilation target flag: Scratch & TurboWarp
