+++ Cargo.toml
@@ -7,11 +7,10 @@ edition = "2024"
 anyhow = "1.0.98"
 itertools = "0.14.0"
 uuid = { version = "1.17.0", features = ["v4"] }
-nom = "8.0.0"
 colored = "3.0.0"
 strum = { version = "0.27.2", features = ["derive"] }
 clap = { version = "4.5.41", features = ["derive"] }
-zip = "3.0.0"                                         # need to use old version for compatibility with zip-extensions
+zip = "3.0.0"                                          # need to use old version for compatibility with zip-extensions
 zip-extensions = "0.8.3"
 fs_extra = "1.3.0"
 serde_json = "1.0.141"
