+++ src/ez/defs.rs
@@ -86,6 +86,7 @@ pub enum ControlOp {
     IfElse { condition: Arc<Expr>, then_substack: Arc<Expr>, else_substack: Arc<Expr> },
     Wait { duration: Arc<Expr> },
     Repeat { times: Arc<Expr>, looped_substack: Arc<Expr> },
+    RepeatUntil { condition: Arc<Expr>, then_substack: Arc<Expr> },
 }
 
 #[derive(Debug)]
@@ -154,7 +155,6 @@ pub enum Literal {
 
 #[derive(Debug)]
 pub struct CustomBlock {
-    pub uuid: Uuid,
     pub name: Arc<str>,
 }
