+++ src/mem_opt/mod.rs
@@ -1,6 +1,5 @@
 pub mod ast;
-pub mod compile_broadcasts;
-pub mod compile_procedures;
+pub mod compile;
 pub mod designate;
 pub mod export;
 pub mod opt;
