+++ src/utils.rs
@@ -8,6 +8,8 @@ use std::{
 use zip::ZipWriter;
 use zip_extensions::ZipWriterExtensions;
 
+use crate::Target;
+
 pub fn time<R>(message: &str, runner: impl FnOnce() -> R) -> R {
     print!("{message} ");
     let start = Instant::now();
@@ -27,13 +29,19 @@ pub fn time_total<R>(message: &str, runner: impl FnOnce() -> R) -> R {
     res
 }
 
-pub fn write_json(json: &str, in_path: &Path, out_path: &Path, res_dir: &include_dir::Dir) {
+pub fn write_json(
+    json: &str,
+    in_path: &Path,
+    out_path: &Path,
+    res_dir: &include_dir::Dir,
+    target: Target,
+) {
     let in_filename = in_path.file_name().and_then(|name| name.to_str());
     let out_name = in_filename.map(|name| name.split('.')).and_then(|mut parts| parts.next());
     let out_name = out_name.expect("input path should have a filename prefix");
 
     let out_folder = Path::new(&out_path).join(out_name);
-    write_files(json, out_name, &out_folder, out_path, res_dir);
+    write_files(json, out_name, &out_folder, out_path, res_dir, target);
 
     let _ = fs::remove_dir_all(&out_folder);
 }
@@ -46,6 +54,7 @@ fn write_files(
     out_folder: &PathBuf,
     out_path: &Path,
     res_dir: &include_dir::Dir,
+    target: Target,
 ) {
     let _ = fs::DirBuilder::new().create(out_folder);
     let out_path_metadata = fs::metadata(out_folder).expect("out path should exist");
@@ -61,7 +70,12 @@ fn write_files(
             .unwrap_or_else(|_| panic!("copying `res/{:?}` should succeed", file.path()));
     }
 
-    let out_zip_path = Path::new(&out_path).join(format!("{out_name}.sb3"));
+    let sub_ext = match target {
+        Target::Scratch => "s",
+        Target::TurboWarp => "t",
+    };
+
+    let out_zip_path = Path::new(&out_path).join(format!("{out_name}.{sub_ext}.sb3"));
     let out_zip_file = File::create(out_zip_path).expect("should be able to create out zip file");
     let zip_writer = ZipWriter::new(out_zip_file);
     zip_writer.create_from_directory(out_folder).expect("creating zip should succeed");
